requirements.txt
requests
python-dotenv
kafka-python
datetime
happybase
pandas
